+++ packages/zenn-cli/package.json
@@ -48,7 +48,6 @@
     "node-fetch": "^2.6.7",
     "open": "^8.2.0",
     "ws": "^7.4.6",
-    "zenn-embed-components": "^0.1.108-alpha.1",
     "zenn-markdown-html": "^0.1.108-alpha.1"
   },
   "devDependencies": {
