--- packages/zenn-embed-components/.eslintrc.js
@@ -1,20 +1,0 @@
-exports.module = {
-  env: {
-    browser: true,
-  },
-
-  extends: [
-    'eslint:recommended',
-    'plugin:@typescript-eslint/eslint-recommended',
-    'plugin:@typescript-eslint/recommended',
-  ],
-
-  parser: ['@typescript-eslint/parser'],
-
-  ignorePatterns: ['lib/**/*.d.ts'],
-
-  rules: {
-    '@typescript-eslint/explicit-module-boundary-types': 'off',
-    '@typescript-eslint/no-explicit-any': 'off',
-  },
-};
