--- packages/zenn-embed-components/.gitignore
@@ -1,1 +1,0 @@
-lib
