--- packages/zenn-embed-components/.storybook/main.js
@@ -1,9 +1,0 @@
-module.exports = {
-  stories: ['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
-  addons: ['@storybook/addon-links', '@storybook/addon-essentials'],
-  framework: '@storybook/react',
-  babel: async (options) => ({
-    ...options,
-    presets: [...options.presets, '@emotion/babel-preset-css-prop'],
-  }),
-};
