--- packages/zenn-embed-components/README.md
@@ -1,3 +1,0 @@
-# zenn-embed-components
-
-> TODO: description
