--- packages/zenn-embed-components/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "zenn-embed-components",
-  "version": "0.1.108-alpha.1",
-  "license": "MIT",
-  "description": "React Component for embed elements in Zenn markdown",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/zenn-dev/zenn-editor.git",
-    "directory": "packages/zenn-embed-components"
-  },
-  "main": "lib/index.js",
-  "files": [
-    "lib"
-  ],
-  "scripts": {
-    "dev": "start-storybook -p 6006",
-    "lint": "eslint . --ext .ts,.tsx",
-    "lint:strict": "eslint . --ext .ts,.tsx --max-warnings 0",
-    "lint:fix": "eslint . --ext .ts,.tsx --fix",
-    "build": "rimraf lib/* && tsc -p ./tsconfig.build.json",
-    "test": "echo 'no test yet'",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
-  },
-  "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "prismjs": "^1.26.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.17.2",
-    "@emotion/babel-preset-css-prop": "^11.2.0",
-    "@storybook/addon-actions": "^6.4.19",
-    "@storybook/addon-essentials": "^6.4.19",
-    "@storybook/addon-links": "^6.4.19",
-    "@storybook/react": "^6.4.19",
-    "@types/mermaid": "^8.2.7",
-    "@types/prismjs": "^1.26.0",
-    "@types/react": "^17.0.39",
-    "@typescript-eslint/eslint-plugin": "^5.11.0",
-    "@typescript-eslint/parser": "^5.11.0",
-    "babel-loader": "^8.2.3",
-    "eslint": "^8.8.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-storybook": "^0.5.6",
-    "mermaid": "^8.14.0",
-    "rimraf": "^3.0.2",
-    "typescript": "^4.5.5"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
