--- packages/zenn-embed-components/src/embed/gist/index.tsx
@@ -1,1 +1,0 @@
-export { EmbedGist } from './EmbedGist';
