--- packages/zenn-embed-components/src/embed/github/EmbedGithubLoading.tsx
@@ -1,19 +1,0 @@
-import { EmbedGithubHeader } from './EmbedGithubHeader';
-import { getGithubLinkInfo } from './utils';
-import { embedGithubStyle, statusMessageStyle } from './styles';
-
-interface EmbedGithubLoadingProps {
-  url?: string;
-}
-
-export const EmbedGithubLoading = ({ url }: EmbedGithubLoadingProps) => {
-  return (
-    <div css={embedGithubStyle}>
-      <EmbedGithubHeader url={url} linkInfo={getGithubLinkInfo(url || '')} />
-
-      <div css={statusMessageStyle}>
-        <p>ローディング中...</p>
-      </div>
-    </div>
-  );
-};
