--- packages/zenn-embed-components/src/embed/github/github.stories.tsx
@@ -1,61 +1,0 @@
-import { ComponentStory, ComponentMeta } from '@storybook/react';
-
-import { EmbedGithub } from './index';
-
-export default {
-  title: 'Embedded/Github',
-  component: EmbedGithub,
-} as ComponentMeta<typeof EmbedGithub>;
-
-const Template: ComponentStory<typeof EmbedGithub> = (args) => (
-  <EmbedGithub {...args} />
-);
-
-export const onPreview = Template.bind({});
-onPreview.args = {
-  src: 'https://github.com/zenn-dev/zenn-editor/blob/canary/lerna.json',
-  content: `{
-  "packages": [
-    "packages/*"
-  ],
-  "version": "0.1.106",
-  "npmClient": "yarn",
-  "useWorkspaces": true
-}`,
-};
-
-export const onLoading = Template.bind({});
-onLoading.args = {
-  isLoading: true,
-};
-
-export const onError = Template.bind({});
-onError.args = {
-  error: new Error(''),
-};
-
-export const onSpecifyingLines = Template.bind({});
-onSpecifyingLines.args = {
-  src: 'https://github.com/zenn-dev/zenn-editor/blob/canary/lerna.json#L3-L5',
-  content: `{
-    "packages": [
-      "packages/*"
-    ],
-    "version": "0.1.106",
-    "npmClient": "yarn",
-    "useWorkspaces": true
-  }`,
-};
-
-export const onSpecifyingStartLine = Template.bind({});
-onSpecifyingStartLine.args = {
-  src: 'https://github.com/zenn-dev/zenn-editor/blob/canary/lerna.json#L5',
-  content: `{
-    "packages": [
-      "packages/*"
-    ],
-    "version": "0.1.106",
-    "npmClient": "yarn",
-    "useWorkspaces": true
-  }`,
-};
