--- packages/zenn-embed-components/src/embed/github/index.tsx
@@ -1,2 +1,0 @@
-export { EmbedGithub } from './EmbedGithub';
-export * from './utils';
