--- packages/zenn-embed-components/src/embed/link-card/index.tsx
@@ -1,1 +1,0 @@
-export * from './EmbedLinkCard';
