--- packages/zenn-embed-components/src/embed/link-card/styles.ts
@@ -1,9 +1,0 @@
-import { css } from '@emotion/react';
-
-export const containerStyles = css`
-  border: solid 1px rgba(92, 147, 187, 0.2);
-  border-radius: 8px;
-  overflow: hidden;
-  font-family: -apple-system, system-ui, BlinkMacSystemFont, 'Segoe UI',
-    'Helvetica Neue', Arial, sans-serif;
-`;
