--- packages/zenn-embed-components/src/embed/mermaid/EmbedMermaidLoading.tsx
@@ -1,7 +1,0 @@
-export const EmbedMermaidLoading = () => {
-  return (
-    <div>
-      <p>Loading...</p>
-    </div>
-  );
-};
