--- packages/zenn-embed-components/src/embed/mermaid/EmbedMermaidNotFound.tsx
@@ -1,7 +1,0 @@
-export const EmbedMermaidNotFound = () => {
-  return (
-    <div>
-      <p>Not Found</p>
-    </div>
-  );
-};
