--- packages/zenn-embed-components/src/embed/mermaid/index.tsx
@@ -1,2 +1,0 @@
-export { EmbedMermaid } from './EmbedMermaid';
-export * from './utils';
