--- packages/zenn-embed-components/src/embed/tweet/EmbedTweetLoading.tsx
@@ -1,7 +1,0 @@
-export const EmbedTweetLoading = () => {
-  return (
-    <div>
-      <p>Loading...</p>
-    </div>
-  );
-};
