--- packages/zenn-embed-components/src/embed/tweet/EmbedTweetNotFound.tsx
@@ -1,7 +1,0 @@
-export const EmbedTweetNotFound = () => {
-  return (
-    <div>
-      <p>Not Found</p>
-    </div>
-  );
-};
