--- packages/zenn-embed-components/src/embed/tweet/index.tsx
@@ -1,1 +1,0 @@
-export { EmbedTweet } from './EmbedTweet';
