--- packages/zenn-embed-components/src/index.ts
@@ -1,7 +1,0 @@
-export * from './embed/github';
-export * from './embed/gist';
-export * from './embed/link-card';
-export * from './embed/mermaid';
-export * from './embed/tweet';
-export * from './embed/types';
-export * from './events/resize';
