--- packages/zenn-embed-components/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ESNext",
-    "module": "ESNext",
-    "jsxImportSource": "@emotion/react",
-    "declaration": true,
-    "jsx": "react-jsx",
-    "outDir": "./lib",
-    "rootDir": "./src",
-    "strict": true,
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true
-  },
-  "exclude": ["node_modules", "lib"],
-  "include": ["**/*.ts", "**/*.tsx"]
-}
